pyproject.toml
[project]
name = "fewpy"
version = "0.1.0"
description = "An Open Source project for Few-Shot Learning"
readme = "README.md"
requires-python = ">=3.9"
authors = ["The Fundação CERTI Organization"]
repository = "https://github.com/fundacaocerti/Fewpy"
license = "MIT"
keywords = [
  "few-shot-learning",
  "machine-learning",
  "meta-learning",
  "artificial-intelligence",
  "data-science",
  "computer-vision",
  "deep-learning",
  "fewpy"
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Operating System :: POSIX :: Linux",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Scientific/Engineering",
  "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
    "albumentations>=2.0.8",
    "clip>=0.2.0",
    "ftfy>=6.3.1",
    "ipykernel>=6.30.1",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pip>=25.2",
    "pydantic>=2.11.7",
    "regex>=2025.7.34",
    "scipy>=1.13.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.7"
]
poe = [
    "poethepoet>=0.36.0",
]

[tool.ruff]
line-length = 80
lint.extend-select = ["I", "W", "E"]
